+++ api/__tests__/fixtures/select-default-cluster.json
@@ -1,8 +1,8 @@
 [
   {
     "id": 1,
-    "name": "kam",
-    "description": "Kamloop cluster",
+    "name": "silver",
+    "description": "Silver cluster",
     "disasterRecovery": false,
     "onPrem": true,
     "createdAt": "2020-09-10T18:14:13.304Z",
